# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Marketplace de Productos para Mascotas API",
        "description": "API RESTful para el marketplace de productos para mascotas. Permite gestionar usuarios, productos, tiendas, carritos y pedidos mediante autenticación Bearer Token (Sanctum).",
        "contact": {
            "name": "Equipo Marketplace Mascotas",
            "email": "admin@marketplace-mascotas.test"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://marketplace_productos_mascotas.test/api",
            "description": "API local"
        }
    ],
    "paths": {
...
        "/orders": {
            "get": {
                "tags": [
                    "Pedidos"
                ],
                "summary": "Ver mis pedidos",
                "description": "Devuelve todos los pedidos del usuario autenticado (rol buyer), con items, pago y entrega.",
                "operationId": "getMyOrders",
                "responses": {
                    "200": {
                        "description": "Lista de pedidos del comprador",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "user_id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "total_amount": {
                                                        "type": "number",
                                                        "format": "float",
                                                        "example": 150
                                                    },
                                                    "status": {
                                                        "type": "string",
                                                        "example": "pending"
                                                    },
                                                    "shipping_address": {
                                                        "type": "string",
                                                        "example": "Calle Falsa 123"
                                                    },
                                                    "notes": {
                                                        "type": "string",
                                                        "example": "Sin notas",
                                                        "nullable": true
                                                    },
                                                    "order_items": {
                                                        "type": "array",
                                                        "items": {
                                                            "type": "object"
                                                        }
                                                    },
                                                    "payment": {
                                                        "type": "object",
                                                        "nullable": true
                                                    },
                                                    "delivery": {
                                                        "type": "object",
                                                        "nullable": true
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2026-04-08T12:00:00.000000Z"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "403": {
                        "description": "Sin permisos"
                    },
                    "500": {
                        "description": "Error interno del servidor"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Pedidos"
                ],
                "summary": "Crear pedido desde el carrito",
                "description": "Convierte el carrito actual en una orden. Descuenta stock, vacía el carrito y retorna la orden creada.",
                "operationId": "storeOrder",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "shipping_address"
                                ],
                                "properties": {
                                    "shipping_address": {
                                        "type": "string",
                                        "example": "Calle Falsa 123"
                                    },
                                    "notes": {
                                        "type": "string",
                                        "example": "Sin notas",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Pedido creado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "total_amount": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 150
                                                },
                                                "status": {
                                                    "type": "string",
                                                    "example": "pending"
                                                },
                                                "shipping_address": {
                                                    "type": "string",
                                                    "example": "Calle Falsa 123"
                                                },
                                                "order_items": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "object"
                                                    }
                                                },
                                                "payment": {
                                                    "type": "object",
                                                    "nullable": true
                                                },
                                                "delivery": {
                                                    "type": "object",
                                                    "nullable": true
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Carrito vacío o stock insuficiente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "El carrito está vacío."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "403": {
                        "description": "Sin permisos"
                    },
                    "500": {
                        "description": "Error interno del servidor"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/orders/{id}": {
            "get": {
                "tags": [
                    "Pedidos"
                ],
                "summary": "Ver detalle de un pedido",
                "description": "Devuelve el detalle completo de un pedido. Los buyers solo pueden ver sus propios pedidos.",
                "operationId": "showOrder",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del pedido",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalle del pedido",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "status": {
                                                    "type": "string",
                                                    "example": "confirmed"
                                                },
                                                "total_amount": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 150
                                                },
                                                "order_items": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "object"
                                                    }
                                                },
                                                "payment": {
                                                    "type": "object",
                                                    "nullable": true
                                                },
                                                "delivery": {
                                                    "type": "object",
                                                    "nullable": true
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Sin permisos para ver este pedido",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "No tienes permiso para esta acción."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "404": {
                        "description": "Pedido no encontrado"
                    },
                    "500": {
                        "description": "Error interno del servidor"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
...
        "/seller/orders": {
            "get": {
                "tags": [
                    "Pedidos"
                ],
                "summary": "Ver pedidos recibidos como vendedor",
                "description": "Devuelve los pedidos que contienen productos del vendedor autenticado.",
                "operationId": "getSellerOrders",
                "responses": {
                    "200": {
                        "description": "Lista de pedidos del vendedor",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "403": {
                        "description": "Sin permisos"
                    },
                    "500": {
                        "description": "Error interno del servidor"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/admin/orders": {
            "get": {
                "tags": [
                    "Pedidos"
                ],
                "summary": "Ver todos los pedidos (admin)",
                "description": "Solo accesible por administradores. Devuelve todos los pedidos del sistema.",
                "operationId": "getAdminOrders",
                "responses": {
                    "200": {
                        "description": "Lista de todos los pedidos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "403": {
                        "description": "Sin permisos de administrador"
                    },
                    "500": {
                        "description": "Error interno del servidor"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "BearerAuth": {
                "type": "http",
                "description": "Ingresá el token obtenido en /api/login. Formato: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}